app/core/requirements.txt
pyopenssl==19.1.0
jinja2==2.11.3
psutil==5.7.3
schedule==0.6.0
flask==1.0.3
paho-mqtt==1.5.0
gunicorn==20.0.4 
setuptools==49.2.1
requests==2.21.0
Werkzeug==0.16.1
simplejson==3.16.0
#backends==0.3.8
urllib3==1.24.3
utils==1.0.1
oauthlib==2.0.1
smbus==1.1.post2
pythondialog==3.5.1
docker==4.4.1
prometheus_client==0.9.0